+++ CHANGELOG.md
@@ -49,6 +49,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Fix getting title field in page mutations #5160 by @maarcingebala
 - Fixed serialization error on weight fields when running `loaddata` and `dumpdb` - #5005 by @NyanKiyoshi
 - Add warehouse model - #4986 by @szewczykmira
+- Assign quantity by warehouse - #4986 by @szewczykmira
 
 ## 2.9.0